app/mcposerver/mcp/mcp-server-demo/pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.40.3",
    "google-cloud-storage>=3.1.1",
    "mcp[cli]>=1.10.1",
    "mcpo>=0.0.15",
    "playwright>=1.53.0",
    "playwright-stealth>=2.0.0",
]
